distributed-docker/datasources/requirements.txt
blinker==1.9.0
boto3==1.35.84
botocore==1.35.84
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.4
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
